Changed default Nixpkgs from unstable to stable.
I was having issues with the unstable branch. gamescope was coredumping
and I couldn't use the version on the stable branch while software that
I used gamescope with was installed from unstable. So I decided to
switch over to the stable branch of Nixpkgs and install packages from
unstable when needed (such as the case with nh, dsda-doom,
dsda-launcher, etc.).