Replaced Nix Colors with Stylix.
Stylix uses a wallpaper to generate a base16 colour scheme. This theme
is then applied to supported programs. The colour of Hyprland's borders
are selected with 'lib.mkForce' to ensure a strong contrast between
active and inactive borders.

The change wallpaper script has been removed in favour of integrating a
wallpaper and theme to a home-manager build. I'm planning of creating a
series of themes and accompanying wallpapers, which can be switched
during runtime.