Added day and night wallpaper switching.
Implemented a script which sets one of two wallpapers depending on
whether it's day or night. A SystemD timer runs a service every 30
minutes, which in turns runs the aforementioned script.

Hyprland didn't seem to run the script successfully on startup. Going by
the script's output in the SystemD service it seems that the Hyprpaper
socket is not created in time for the script to run. This is remedied by
adding a 3 second delay before the script runs, in which a greyscale
version of the day wallpaper variant is shown in the meantime.

Version 0.7.0 of Hyprpaper from the unstable branch of nixpkgs is being
used since a change to socket location in hyprpaper means that IPC does
not work on 0.6.0, the version currently in the stable branch.
https://github.com/hyprwm/hyprpaper/issues/5#issuecomment-2120838303