Added script to toggle laptop display.
Added a script which toggles my laptops display. This is useful for when
I only want an external display on. The 'toggle' argument of the dpms
dispatcher was not working, so a JSON query for display status followed
by an if-else branch is used instead.

The one second sleep before dispatch was added to avoid potential
undefined behaviour, as stated in Hyprland's wiki.
https://wiki.hyprland.org/Configuring/Dispatchers/#list-of-dispatchers