Added battery scripts to replace Waybar modules.
Replaced the individual battery modules in Waybar with a single custom
module that calls a script to get the status of each battery connected
to the system. This makes the display of battery information extensible
as we no longer need a separate module for each battery.

Clicking on the module will display a notification containing more
detailed information about each battery.