Replaced Physlock with Hyprlock.
When resuming from a suspend, Physlock would sometimes fail to show its
prompt or would seemingly crash Hyprland. For these reasons I have
switched to Hyprlock.

However, this does introduce the security risk of someone switching to a
already logged-in TTY. This risk can be mitigated by reducing the number
of accessable TTYs.