Revert "Updated system."
This reverts commit 68ed35e607c172e66d613a165eb4a33c57cf0e39.

Both hyprpaper and xdg-desktop-portal-hyprland fail to build. Pull
requests to fix these have been merged but have yet to be included in
nixos-unstable.

I will refrain from updating until these are included. The stable
version of xdg-desktop-portal-hyprland has an RCE whose patch is not
included in the derivation and the stable version of hyprpaper does not
seem to open its socket.