Modified MangoHud to use mkOutOfStoreSymlink.
mkOutOfStoreSymlink allows a file to be symlinked outside of the Nix
Store, allowing said file to be modified without having to rebuild the
system to see the changes [1]. Thanks to nanoni17728 for explaining that
the path to the config file has to be absolute [2].

I chose MangoHud since I may want to reduce the FPS cap on games that
have issues running at high framerates, and the built-in cap is not
ideal. While it's possible to set the FPS cap with MANGOHUD_CONFIG, its
options take priority over every other config file [3], effectively
overriding MangoHud.conf's options.

[1] https://nixos-and-flakes.thiscute.world/best-practices/accelerating-dotfiles-debugging
[2] https://github.com/ryan4yin/nixos-and-flakes-book/discussions/168#discussioncomment-11811475
[3] https://github.com/flightlessmango/MangoHud?tab=readme-ov-file#environment-variables