Forced home-pi to use 5GHz WiFi by setting BSSID.
'freq_list' was dropping connection presumably when the router changed
WiFi channel. After reading a post on Snapcraft suggesting to connect
via BSSID instead, I decided to give that a try. [1]

I replaced the SSID secret with one titled 'BSSID' which uses the BSSID
of the 5GHz band. The PSK has to be a base64 encoded password derived
from both the SSID and ASCII passsword, which can be calculated using
the command 'wpa_passphrase <SSID> <password>'. [2]

[1] https://forum.snapcraft.io/t/provide-wpa-supplicant-conf-to-networkmanager-to-disable-5ghz/29598
[2] https://raspberrypi.stackexchange.com/a/139389